Safari-609.1.11.1/LayoutTests/js/date-daysfrom1970-overflow-expected.txt

This test confirms an assertion in dateToDaysFrom1970() in wtf/DateMath.cpp passes. The function had a bug setting a number larger than INT_MAX to an int variable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d.getTime() is NaN
PASS Date.UTC(1970, 0, 1, 0, 0, 0, 0) is 0
PASS d.getTime() is NaN
PASS successfullyParsed is true

TEST COMPLETE
